+++ lib/util/collision_game_component.dart
@@ -1,0 +1,35 @@
+import 'package:bonfire/base/game_component.dart';
+import 'package:bonfire/bonfire.dart';
+import 'package:bonfire/collision/object_collision.dart';
+import 'package:flame/extensions.dart';
+
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 10/12/21
+
+class CollisionGameComponent extends GameComponent with ObjectCollision {
+  final String name;
+  final Map<String, dynamic>? properties;
+
+  CollisionGameComponent({
+    this.name = '',
+    this.properties,
+    required Vector2 position,
+    List<CollisionArea> collisions = const [],
+  }) {
+    this.position = this.position.copyWith(
+          position: position,
+        );
+    setupCollision(
+      CollisionConfig(collisions: collisions),
+    );
+  }
+}